requirements.txt
fastapi
uvicorn[standard]
uvloop
watchfiles
websockets
twilio
deepgram-sdk>=3.0.0,<4.0.0
google-genai
httpx
python-dotenv
pydantic
pyjwt
passlib
bcrypt==3.2.2
python-multipart
chromadb
pymupdf
numpy
groq
faiss-cpu
sentence-transformers
redis
pymysql
DBUtils
grpcio>=1.62.0
grpcio-tools>=1.62.0
protobuf>=4.25.0
